requirements.txt
strands-agents>=1.0.0
strands-agents-tools>=0.2.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
gspread>=5.0.0
pillow>=9.0.0
openai>=1.0.0
python-telegram-bot>=20.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0